+++ hpyc_pyside_ui/__main__.py
@@ -5,6 +5,7 @@ import dbm
 import sys
 
 import hpyculator
+import hpyc_core
 
 # from PySide6 import QtWidgets, QtCore
 from PySide6.QtWidgets import QApplication