+++ hpyc_pyside_ui/utils/var.py
@@ -1,0 +1,3 @@
+from hpyc_core import Core
+
+instance_core = Core()